# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.DataTypeRequest:
    properties:
      data_type:
        type: string
    required:
    - data_type
    type: object
...
  handlers.FormGroupRequest:
    properties:
      group_name:
        type: string
      group_row:
        type: integer
      group_span:
        type: integer
    type: object
...
  structs.ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: string
      status:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the Kora API server for form management.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Kora API
  version: "1.0"
paths:
...
  /data_types:
    get:
      consumes:
      - application/json
      description: Retrieve all data types
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      summary: Get all data types
      tags:
      - data-types
    post:
      consumes:
      - application/json
      description: Create a new data type
      parameters:
      - description: Data Type Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.DataTypeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      summary: Create data type
      tags:
      - data-types
  /data_types/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a data type by its ID
      parameters:
      - description: Data Type ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      summary: Delete data type
      tags:
      - data-types
    get:
      consumes:
      - application/json
      description: Retrieve a data type by its ID
      parameters:
      - description: Data Type ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      summary: Get data type
      tags:
      - data-types
    put:
      consumes:
      - application/json
      description: Update an existing data type by its ID
      parameters:
      - description: Data Type ID
        in: path
        name: id
        required: true
        type: integer
      - description: Data Type Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.DataTypeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      summary: Update data type
      tags:
      - data-types
...
  /form_groups:
    get:
      consumes:
      - application/json
      description: Retrieve all form groups
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      summary: Get all form groups
      tags:
      - form-groups
    post:
      consumes:
      - application/json
      description: Create a new form group
      parameters:
      - description: Form Group Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.FormGroupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      summary: Create form group
      tags:
      - form-groups
  /form_groups/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a form group by its ID
      parameters:
      - description: Form Group ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      summary: Delete form group
      tags:
      - form-groups
    get:
      consumes:
      - application/json
      description: Retrieve a form group by its ID
      parameters:
      - description: Form Group ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      summary: Get form group
      tags:
      - form-groups
    put:
      consumes:
      - application/json
      description: Update an existing form group by its ID
      parameters:
      - description: Form Group ID
        in: path
        name: id
        required: true
        type: integer
      - description: Form Group Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.FormGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      summary: Update form group
      tags:
      - form-groups
...
schemes:
- http
- https
swagger: "2.0"